# perf: plugin scan perf optimization