--- a/packages/amplify-cli/src/plugin-manager.ts
+++ b/packages/amplify-cli/src/plugin-manager.ts
@@ -7,7 +7,7 @@ import {
   getCorePluginVersion,
   isUnderScanCoverageSync,
 } from './plugin-helpers/scan-plugin-platform';
-import { verifyPlugin, verifyPluginSync } from './plugin-helpers/verify-plugin';
+import { verifyPlugin } from './plugin-helpers/verify-plugin';
 import createNewPlugin from './plugin-helpers/create-new-plugin';
 import { AddPluginResult, AddPluginError } from './domain/add-plugin-result';
 import { twoPluginsAreTheSame } from './plugin-helpers/compare-plugins';
@@ -154,16 +154,16 @@ export async function confirmAndScan(pluginPlatform: PluginPlatform) {
   }
 }
 
-export function addUserPluginPackage(pluginPlatform: PluginPlatform, pluginDirPath: string): AddPluginResult {
+export const addUserPluginPackage = async (pluginPlatform: PluginPlatform, pluginDirPath: string): Promise<AddPluginResult> => {
   return addPluginPackage(pluginPlatform, pluginDirPath);
-}
+};
 
-export function addExcludedPluginPackage(pluginPlatform: PluginPlatform, pluginInfo: PluginInfo): AddPluginResult {
+export const addExcludedPluginPackage = async (pluginPlatform: PluginPlatform, pluginInfo: PluginInfo): Promise<AddPluginResult> => {
   return addPluginPackage(pluginPlatform, pluginInfo.packageLocation);
-}
+};
 
-export function addPluginPackage(pluginPlatform: PluginPlatform, pluginDirPath: string): AddPluginResult {
-  const pluginVerificationResult = verifyPluginSync(pluginDirPath);
+export const addPluginPackage = async (pluginPlatform: PluginPlatform, pluginDirPath: string): Promise<AddPluginResult> => {
+  const pluginVerificationResult = await verifyPlugin(pluginDirPath);
   const result = new AddPluginResult(false, pluginVerificationResult);
 
   if (pluginVerificationResult.verified) {
@@ -209,7 +209,7 @@ export function addPluginPackage(pluginPlatform: PluginPlatform, pluginDirPath: 
     result.error = AddPluginError.FailedVerification;
   }
   return result;
-}
+};
 
 // remove: select from the plugins only,
 // if the location belongs to the scan directories, put the info inside the excluded.