*Description of changes:*

- fix: optimize plugin scan performance, to lazy load plugins
    before:
        ```amplify plugin scan  5.86s user 1.58s system 89% cpu 8.327 total```
    after:
        ```amplify-dev plugin scan  0.68s user 0.12s system 116% cpu 0.684 total```
- remove extra verification functions


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.